make clippy happy

+++ src/node_48.rs
@@ -98,7 +98,7 @@ impl Node for Node48 {
 
     fn insert(&mut self, key: u8, node: NodePtr) {
         let pos = self.next_empty as usize;
-        self.next_empty = self.children[pos as usize].as_tid() as u8;
+        self.next_empty = self.children[pos].as_tid() as u8;
 
         debug_assert!(pos < 48);
